+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- [ISSUE-119](https://github.com/dailymotion/tartiflette/issues/119): Add an `initial_value` parameter on the `execute`'s engine method.
+
 ## [Released]
 
 ## [0.3.7] - 2019-02-06